requirements.txt
fastapi
uvicorn
pandas
google-cloud-storage
google-cloud-bigquery
